refactor: move WeatherBackend protocol and related models to `weather\backends\protocol.py`
